Tool class - contains tool, exporter and builder - all templates for a tool
