+++ project_generator/tools/eclipse.py
@@ -18,12 +18,11 @@ import logging
 # eclipse works with linux paths
 from posixpath import normpath, join, basename
 
-from .exporter import Exporter
-from .builder import Builder
+from .tool import Tool, Builder, Exporte
 from .gccarm import MakefileGccArm
 
 
-class EclipseGnuARM(Exporter, Builder):
+class EclipseGnuARM(Tool, Exporter, Builder):
     source_files_dic = ['source_files_c', 'source_files_s',
                         'source_files_cpp', 'source_files_obj']
     file_types = {'cpp': 1, 'c': 1, 's': 1, 'obj': 1, 'lib': 1}
