+++ project_generator/tools/gdb.py
@@ -6,9 +6,7 @@
 
 import copy
 
-from .exporter import Exporter
-from .builder import Builder
-from .tool import Tool
+from .tool import Tool, Builder, Exporte
 
 class gdb_definitions():
